Load the ui token from env var, if required
